Unix socket support
The new Brave BAT payments service uses viproxy to expose the service in
the nitro enclave. For local testing outside AWS we plan to simulate
Nitro by using a container that does not have an IP network but rather
uses Unix sockets to communicate to the outside world. To cover that add
support for directing the traffic into a unix socket.

For symmetry also allow to listen on UNIX sockets.

Report an explicit error when net.Addr is unsupported.